+++ config/config_sample.js
@@ -11,7 +11,14 @@ module.exports = {
   password: null,             // change it to set a server password for user access
   shutdownDelay: 0,           // delay before server kill on SIGTERM/SIGINT - useful in production
 
-  locales: ["en","fr", "es"],
-  defaultLocale: "en"
+  maxRooms: 10,               // maximum number of rooms allowed
+  maxPlayers: 100,            // maximum number of connected players allowed
 
-}
+  locales: ["en","fr", "es"],
+  defaultLocale: "en",
+
+  name: "",                   // custom server name
+  welcomeMsg: "",             // shown on "home" page
+  motd: "",                   // shown on "lobby" page
+  description: "",            // shown on "about" page
+};