+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+v0.2.0
+------
+2016-02-26
+
+- Add username save in URL
+  + Useful when refreshing the page
+  + Useful when using bookmarks
+- Add sound support (#23)
+  + Add new events to handle sounds from modules
+  + Add mute/Unmute button available ingame. Preferences are saved in browser.
+- Add resize option in Pre-game Chat (#29)
+- Add some quotas for maximum number of players and room (#26)
+- Add custom welcome message and message of the day (motd)
+- Fix visibility of "Leave Room" button ingame
+- Ensure compatibility with latest version of node and socket.io
+
 v0.1.3
 ------
 2015-08-26